// src/lib/apex/types.ts
export type ApexContractId = "UNDER6" | "UNDER7" | "UNDER8" | "OVER1" | "OVER2" | "OVER3";

export const APEX_CONTRACTS: ApexContractId[] = [
  "UNDER7",
  "OVER2",
  "UNDER6",
  "UNDER8",
  "OVER1",
  "OVER3",
];